Vertex Arrays, Buffers, Index Buffers
Working triangles and runtime defined buffer layouts

+++ Neurex/src/Neurex/imgui/ImGuiLayer.cpp
@@ -86,7 +86,7 @@ namespace Neurex {
 			ImGuiIO& io = ImGui::GetIO();
 			io.DisplaySize = ImVec2(e.get_width(), e.get_height());
 			io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
-			glViewport(0, 0, e.get_width(), e.get_height());
+			glViewport(0, 0, (int)e.get_width(), (int)e.get_height());
 			return false;
 		});
